backend/services/agent-orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]>=0.30.6
python-dotenv>=1.0.1
openai>=1.40.0
crewai==1.9.3
crewai-tools==1.9.3
azure-ai-inference>=1.0.0b9
sse-starlette==2.1.3
redis==5.0.8
pydantic>=2.8.2
httpx>=0.27.0
python-multipart==0.0.9
wikipedia-api>=0.6.0
